+++ src/hints/Hints.ts
@@ -61,11 +61,3 @@ export function decodeHint(hint: Hint): DecodedHint {
             };
     }
 }
-
-export default function HintDescription({ hint, area }: { hint: DecodedHint, area?: string }) {
-    return (
-        <div style={{ color: `var(--scheme-${hint.style})` }}>
-            {area && `${area} - `}{hint.description}
-        </div>
-    );
-}
