+++ src/hints/HintsParser.ts
@@ -1,7 +1,7 @@
-import { decodeHint, type Hint } from '../locationTracker/Hints';
 import dungeonData from '../data/dungeons.json';
 import { itemMaxes } from '../logic/Inventory';
 import type { SubmarkerData } from '../locationTracker/mapTracker/Marker';
+import { decodeHint, type Hint } from './Hints';
 
 const barrenSots = /^([a-z0-9\s']+) (barren|dead|sots)$/;
 const path = /^([a-z0-9\s']+) -> ([a-z0-9\s']+)\s*(?:\(.*\))?$/;
