+++ src/locationTracker/HintsDescription.tsx
@@ -1,0 +1,16 @@
+import type { DecodedHint } from '../hints/Hints';
+
+export default function HintDescription({
+    hint,
+    area,
+}: {
+    hint: DecodedHint;
+    area?: string;
+}) {
+    return (
+        <div style={{ color: `var(--scheme-${hint.style})` }}>
+            {area && `${area} - `}
+            {hint.description}
+        </div>
+    );
+}
